Renamed personalCode to personalCodeHandler
